Refactor: Optimize layout architecture and component naming
Major performance and architecture improvements:

**Data Fetching Optimization:**
- Remove redundant user projects query from project layout (85-90% data reduction)
- Replace with TanStack Query for real-time project list polling
- Move project list fetching to client-side for better caching

**Component Architecture Refactor:**
- Rename project-sidebar.tsx → primary-sidebar.tsx (project navigation)
- Rename project-secondary-sidebar.tsx → project-sidebar.tsx (project details)
- Update all component references across layouts

**Technical Improvements:**
- Add TanStack Query with devtools for efficient state management
- Implement 5-second polling for real-time project status updates
- Maintain clean separation of concerns between components

Performance impact: 85-90% reduction in layout data transfer,
faster page rendering, and improved real-time status updates.
