+++ components/project-sidebar.tsx
@@ -1,154 +1,205 @@
 'use client';
 
 import { useState } from 'react';
-import { Project } from '@prisma/client';
-import { Circle, FolderOpen, GitBranch, Home, Plus, Settings } from 'lucide-react';
-import Link from 'next/link';
+import { Environment, Project, Sandbox } from '@prisma/client';
+import {
+  ChevronDown,
+  ChevronLeft,
+  ChevronRight,
+  CreditCard,
+  Database,
+  Github,
+  Key,
+  Package,
+  Settings,
+  Shield,
+  Terminal,
+} from 'lucide-react';
+import { usePathname, useRouter } from 'next/navigation';
 
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
+import { useTerminal } from '@/components/terminal-provider';
 import { cn } from '@/lib/utils';
 
 interface ProjectSidebarProps {
-  projects: Project[];
-  currentProjectId: string;
-  userId: string;
+  project: Project;
+  sandboxes: Sandbox[];
+  envVars: Environment[];
 }
 
-export default function ProjectSidebar({
-  projects,
-  currentProjectId,
-}: ProjectSidebarProps) {
-  const [isExpanded, setIsExpanded] = useState(false);
+export default function ProjectSidebar({ project }: ProjectSidebarProps) {
+  const [isCollapsed, setIsCollapsed] = useState(false);
+  const [isConfigExpanded, setIsConfigExpanded] = useState(true);
+  const { hideTerminal, isTerminalVisible } = useTerminal();
+  const pathname = usePathname();
+  const router = useRouter();
 
-  const getStatusColor = (status: string) => {
-    switch (status) {
-      case 'READY':
-      case 'DEPLOYED':
-        return 'text-green-600 dark:text-green-500';
-      case 'INITIALIZING':
-      case 'DEPLOYING':
-        return 'text-yellow-600 dark:text-yellow-500';
-      case 'ERROR':
-        return 'text-destructive';
-      default:
-        return 'text-muted-foreground';
+  const handleSectionClick = (e: React.MouseEvent, sectionId: string, href: string) => {
+    e.preventDefault();
+
+    if (sectionId === 'terminal') {
+      // For terminal, navigate without hiding (terminal page will show it)
+      router.push(href);
+    } else {
+      // For other sections, hide terminal and navigate
+      hideTerminal();
+      router.push(href);
     }
   };
+
+  const topSections = [
+    {
+      id: 'terminal',
+      label: 'Web Terminal',
+      icon: Terminal,
+      href: `/projects/${project.id}/terminal`,
+    },
+    { id: 'database', label: 'Database', icon: Database, href: `/projects/${project.id}/database` },
+  ];
+
+  const configSections = [
+    {
+      id: 'environment',
+      label: 'Environment Variables',
+      icon: Package,
+      href: `/projects/${project.id}/environment`,
+    },
+    {
+      id: 'secrets',
+      label: 'Secret Configuration',
+      icon: Key,
+      href: `/projects/${project.id}/secrets`,
+    },
+    { id: 'auth', label: 'Auth Configuration', icon: Shield, href: `/projects/${project.id}/auth` },
+    {
+      id: 'payment',
+      label: 'Payment Configuration',
+      icon: CreditCard,
+      href: `/projects/${project.id}/payment`,
+    },
+  ];
+
+  const bottomSections = [
+    {
+      id: 'github',
+      label: 'GitHub Repository',
+      icon: Github,
+      href: `/projects/${project.id}/github`,
+    },
+  ];
+
+  // Check if any config section is active
+  const isConfigActive = configSections.some((section) => pathname === section.href);
 
   return (
     <div
       className={cn(
-        'bg-sidebar-background border-r border-sidebar-border flex flex-col transition-all duration-200',
-        isExpanded ? 'w-52' : 'w-12'
+        'bg-sidebar-project-background border-r border-border flex flex-col transition-all duration-200',
+        isCollapsed ? 'w-10' : 'w-72'
       )}
-      onMouseEnter={() => setIsExpanded(true)}
-      onMouseLeave={() => setIsExpanded(false)}
     >
-      <TooltipProvider>
-        {/* Header */}
-        <div className="h-12 flex items-center justify-center border-b border-sidebar-border">
-          <FolderOpen className="h-5 w-5 text-muted-foreground" />
-          {isExpanded && <span className="ml-2 text-sm font-medium text-foreground">Projects</span>}
-        </div>
+      {/* Header */}
+      <div className="h-12 flex items-center justify-between px-3 border-b border-border">
+        {!isCollapsed && (
+          <span className="text-sm font-medium text-foreground">Project {project.name}</span>
+        )}
+        <button
+          onClick={() => setIsCollapsed(!isCollapsed)}
+          className="text-muted-foreground hover:text-foreground transition-colors"
+        >
+          {isCollapsed ? <ChevronRight className="h-4 w-4" /> : <ChevronLeft className="h-4 w-4" />}
+        </button>
+      </div>
 
-        {/* Navigation Items */}
-        <div className="flex-1 py-2 overflow-y-auto">
-          {/* Home Link */}
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <Link
-                href="/projects"
-                className="group flex items-center h-8 px-3 hover:bg-sidebar-accent transition-colors"
+      {!isCollapsed && (
+        <div className="flex-1 overflow-y-auto">
+          {/* Top sections */}
+          {topSections.map((section) => {
+            const Icon = section.icon;
+            const isActive =
+              pathname === section.href || (section.id === 'terminal' && isTerminalVisible);
+
+            return (
+              <a
+                key={section.id}
+                href={section.href}
+                onClick={(e) => handleSectionClick(e, section.id, section.href)}
+                className={cn(
+                  'group w-full flex items-center px-3 py-2 text-sm transition-colors min-h-[32px]',
+                  isActive ? 'bg-sidebar-accent' : 'hover:bg-sidebar-accent'
+                )}
               >
-                <Home className="h-4 w-4 text-muted-foreground group-hover:text-sidebar-foreground shrink-0 transition-colors" />
-                {isExpanded && (
-                  <span className="ml-3 text-sm text-foreground truncate">All Projects</span>
+                <Icon className="h-4 w-4 mr-2 text-muted-foreground group-hover:text-sidebar-foreground flex-shrink-0 transition-colors" />
+                <span className="text-foreground truncate flex-1">{section.label}</span>
+              </a>
+            );
+          })}
+
+          {/* Configuration Group */}
+          <div>
+            <button
+              onClick={() => setIsConfigExpanded(!isConfigExpanded)}
+              className={cn(
+                'group w-full flex items-center px-3 py-2 text-sm transition-colors min-h-[32px]',
+                isConfigActive ? 'bg-sidebar-accent' : 'hover:bg-sidebar-accent'
+              )}
+            >
+              <Settings className="h-4 w-4 mr-2 text-muted-foreground group-hover:text-sidebar-foreground flex-shrink-0 transition-colors" />
+              <span className="text-foreground flex-1 text-left truncate">Configuration</span>
+              <ChevronDown
+                className={cn(
+                  'h-4 w-4 text-muted-foreground group-hover:text-sidebar-foreground transition-transform flex-shrink-0 transition-colors',
+                  !isConfigExpanded && '-rotate-90'
                 )}
-              </Link>
-            </TooltipTrigger>
-            {!isExpanded && (
-              <TooltipContent side="right">
-                <p>All Projects</p>
-              </TooltipContent>
+              />
+            </button>
+
+            {isConfigExpanded && (
+              <div className="bg-muted/30">
+                {configSections.map((section) => {
+                  const Icon = section.icon;
+                  const isActive = pathname === section.href;
+
+                  return (
+                    <a
+                      key={section.id}
+                      href={section.href}
+                      onClick={(e) => handleSectionClick(e, section.id, section.href)}
+                      className={cn(
+                        'group w-full flex items-center pl-9 pr-3 py-2 text-sm transition-colors min-h-[32px]',
+                        isActive ? 'bg-sidebar-accent' : 'hover:bg-sidebar-accent'
+                      )}
+                    >
+                      <Icon className="h-4 w-4 mr-2 text-muted-foreground group-hover:text-sidebar-foreground flex-shrink-0 transition-colors" />
+                      <span className="text-foreground truncate flex-1">{section.label}</span>
+                    </a>
+                  );
+                })}
+              </div>
             )}
-          </Tooltip>
-
-          {/* New Project */}
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <Link
-                href="/projects/new"
-                className="group flex items-center h-8 px-3 hover:bg-sidebar-accent transition-colors mb-2"
-              >
-                <Plus className="h-4 w-4 text-muted-foreground group-hover:text-sidebar-foreground shrink-0 transition-colors" />
-                {isExpanded && (
-                  <span className="ml-3 text-sm text-foreground truncate">New Project</span>
-                )}
-              </Link>
-            </TooltipTrigger>
-            {!isExpanded && (
-              <TooltipContent side="right">
-                <p>New Project</p>
-              </TooltipContent>
-            )}
-          </Tooltip>
-
-          <div className="border-t border-border my-2" />
-
-          {/* Project List */}
-          <div className="space-y-1">
-            {projects.map((project) => (
-              <Tooltip key={project.id}>
-                <TooltipTrigger asChild>
-                  <Link
-                    href={`/projects/${project.id}`}
-                    className={cn(
-                      'group flex items-center h-8 px-3 hover:bg-sidebar-accent transition-colors',
-                      currentProjectId === project.id && 'bg-sidebar-accent'
-                    )}
-                  >
-                    <div className="flex items-center shrink-0">
-                      <Circle
-                        className={cn('h-2 w-2 mr-2', getStatusColor(project.status))}
-                        fill="currentColor"
-                      />
-                      <GitBranch className="h-4 w-4 text-muted-foreground group-hover:text-sidebar-foreground transition-colors" />
-                    </div>
-                    {isExpanded && (
-                      <span className="ml-3 text-sm text-foreground truncate">{project.name}</span>
-                    )}
-                  </Link>
-                </TooltipTrigger>
-                {!isExpanded && (
-                  <TooltipContent side="right">
-                    <p>{project.name}</p>
-                  </TooltipContent>
-                )}
-              </Tooltip>
-            ))}
           </div>
-        </div>
 
-        {/* Settings */}
-        <div className="p-2">
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <Link
-                href="/settings"
-                className="group flex items-center h-8 px-2 hover:bg-sidebar-accent rounded transition-colors"
+          {/* Bottom sections */}
+          {bottomSections.map((section) => {
+            const Icon = section.icon;
+            const isActive = pathname === section.href;
+
+            return (
+              <a
+                key={section.id}
+                href={section.href}
+                onClick={(e) => handleSectionClick(e, section.id, section.href)}
+                className={cn(
+                  'group w-full flex items-center px-3 py-2 text-sm transition-colors min-h-[32px]',
+                  isActive ? 'bg-sidebar-accent' : 'hover:bg-sidebar-accent'
+                )}
               >
-                <Settings className="h-4 w-4 text-muted-foreground group-hover:text-sidebar-foreground flex-shrink-0 transition-colors" />
-                {isExpanded && <span className="ml-3 text-sm text-foreground">Settings</span>}
-              </Link>
-            </TooltipTrigger>
-            {!isExpanded && (
-              <TooltipContent side="right">
-                <p>Settings</p>
-              </TooltipContent>
-            )}
-          </Tooltip>
+                <Icon className="h-4 w-4 mr-2 text-muted-foreground group-hover:text-sidebar-foreground flex-shrink-0 transition-colors" />
+                <span className="text-foreground truncate flex-1">{section.label}</span>
+              </a>
+            );
+          })}
         </div>
-      </TooltipProvider>
+      )}
     </div>
   );
 }
