+++ package.json
@@ -25,6 +25,8 @@
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-tabs": "^1.1.13",
     "@radix-ui/react-tooltip": "^1.2.8",
+    "@tanstack/react-query": "^5.90.7",
+    "@tanstack/react-query-devtools": "^5.90.2",
     "@zjy365/sealos-desktop-sdk": "^0.1.20",
     "bcryptjs": "^3.0.2",
     "class-variance-authority": "^0.7.1",
